refactor: compartment new component FavoriteButton
