+++ src/components/MovieCard/index.tsx
@@ -1,39 +1,20 @@
 import { Link } from 'react-router-dom';
-import { useSelector } from 'react-redux';
-import useActions from '../../hooks/useActions';
-import { RootState } from '../../store/store';
 import { Movie } from '../../utils/apiResponse';
+import FavoriteButton from '../FavoriteButton';
 
 import noMovieImg from '../../assets/no-image.svg';
-import favMovie from '../../assets/fav.svg';
-import noFavMovie from '../../assets/no-fav.svg';
 import './style.scss';
 
-const MovieCard = ({ data }: { data: Movie[] }) => {
-  const { toggleSelected } = useActions();
-  const selected = useSelector((state: RootState) => state.selected);
-  const isExist = (item: string) => selected.includes(item);
+const MovieCard = ({ movieData }: { movieData: Movie[] }) => {
+  if (movieData.length === 0) return <h3>No Movies</h3>;
 
   return (
     <div className="movie__wrapper">
-      {data &&
-        data.map((movie: Movie) => (
+      {movieData &&
+        movieData.map((movie: Movie) => (
           <div className="movie__item" key={movie.imdbID}>
-            <button
-              type="button"
-              className="movie__button button"
-              onClick={() => toggleSelected(movie.imdbID)}
-            >
-              <img
-                src={isExist(movie.imdbID) ? favMovie : noFavMovie}
-                alt={
-                  isExist(movie.imdbID)
-                    ? 'Favorite movie'
-                    : 'Not favorite movie'
-                }
-              />
-            </button>
-            <Link to={`movie/${movie.imdbID}`}>
+            <FavoriteButton imdbID={movie.imdbID} />
+            <Link to={`/movie/${movie.imdbID}`}>
               <img
                 className="movie__img"
                 src={movie.Poster === 'N/A' ? noMovieImg : movie.Poster}
